refactor : CoinController의 패키지 변경사항
